requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.11
google-generativeai==0.4.1
google-api-core==2.25.1
google-auth==2.40.3
protobuf==4.25.8
faiss-cpu>=1.9.0
requests==2.31.0